requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
asgiref==3.8.1
astrapy==1.5.2
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==3.1.0
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_resources==6.4.5
ipywidgets==8.1.5
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kubernetes==31.0.0
langchain==0.3.8
langchain-anthropic==0.3.0
langchain-chroma==0.1.4
langchain-community==0.3.7
langchain-core==0.3.21
langchain-huggingface==0.1.2
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.143
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.15.1
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.1
onnxruntime==1.20.1
openai==1.54.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.11
overrides==7.7.0
pandas==2.2.3
pillow==11.0.0
posthog==3.7.4
propcache==0.2.0
protobuf==5.29.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pymongo==4.10.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dotenv==1.0.1
pytz==2024.2
#pywin32
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.4
toml==0.10.2
tomli==2.2.1
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typer==0.13.1
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uuid6==2024.7.10
uvicorn==0.32.1
watchdog==6.0.0
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xxhash==3.5.0
yarl==1.17.2
